messages service was modified (user id was extracted from dto to method signature)

+++ Messager.Chats.Infrastructure.Services/Services/MessagesService.cs
@@ -5,6 +5,7 @@ using Messager.Chats.Domain.Core.Models;
 using Messager.Chats.Domain.Interfaces.Repositories;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace Messager.Chats.Infrastructure.Services.Services
@@ -20,19 +21,22 @@ namespace Messager.Chats.Infrastructure.Services.Services
             _mapper = mapper;
         }
 
-        public async Task<MessageForReadDto> CreateMessageAsync(MessageForCreateDto messageDto)
+        public async Task<MessageCreatedForReadDto> CreateMessageAsync(Guid userId, MessageForCreateDto messageDto)
         {
             var message = _mapper.Map<Message>(messageDto);
             message.SendingTime = DateTime.Now;
+            message.SenderId = userId;
             await _repositoryManager.Messages.CreateMessageAsync(message);
             await _repositoryManager.SaveAsync();
-            var messageCreatedDto = _mapper.Map<MessageForReadDto>(message);
+            var messageCreatedDto = _mapper.Map<MessageCreatedForReadDto>(message);
             return messageCreatedDto;
         }
 
-        public async Task DeleteMessageByIdAsync(Guid messageId)
+        public async Task DeleteMessageByIdAsync(Guid userId, Guid messageId)
         {
             var message = await _repositoryManager.Messages.GetMessageByIdAsync(messageId, false);
+            if (!message.SenderId.Equals(userId))
+                throw new InvalidOperationException("invalid message to be deleted");
             _repositoryManager.Messages.DeleteMessage(message);
             await _repositoryManager.SaveAsync();
         }
@@ -58,10 +62,11 @@ namespace Messager.Chats.Infrastructure.Services.Services
             return messageDto;
         }
 
-        public void UpdateMessage(Guid messageId, MessageForUpdateDto messageDto)
+        public void UpdateMessage(Guid userId, Guid messageId, MessageForUpdateDto messageDto)//TODO: check message userId 
         {
             var message = _mapper.Map<Message>(messageDto);
             message.Id = messageId;
+            message.SenderId = userId;
             _repositoryManager.Messages.UpdateMessage(message);
             _repositoryManager.SaveAsync();
         }
